+++ tests/fixtures/issue-1373-merged-namespace-star-reexport/src/merged.ts
@@ -1,0 +1,6 @@
+export declare namespace Merged {
+  type Foo = string;
+}
+
+export const Merged = Object.assign(function Merged() {}, { bar: 1 });
+export const unusedControl = 1;